requirements.txt
discord.py
discord-pretty-help
DiscordUtils